Add 'myheadfoot' package
'myheadfoot' package redefine headers and footers for document pages.
A central page number has been used on all pages.
Even pages has the current chapter title on head.
Odd pages has the current section title on head.
A improvement has been done in 'mytodo' package to mark the title of 'List of "todos"' in order to have it on headers.