Add '.vimrc' for each file
Change the way default Vim configuration are read.
Now, a default Vim configuration can be made for all files and for each files.
'.vimrc' will be the global configuration for all file open in the directory.
'.vimrc.main.tex' will be the configuration file for the file 'main.tex'.
Need to put additional code in the file '~/.vimrc'.