Improve 'mytodo' commands
'\mytodo' and environment 'myTodo' was very visually agressives.
Particularly 'myTodo' environment which introduce new text in the document.
Now, all additionnal text is in the margin.
'myTodo' environment change text color of the document in gray, '\mytodo' command change the color to 'myred'.
Moreover, a vertical line along the concerned lines is drawn in 'myred' color.