Fix '\labelformat' for 'subfigure' and 'subtable'
When 'subfigure' or 'subtable' elements where called, '\labelformat' does not recognize them as a 'figure' or 'table' and produce the default value for reference without prefix '\figurename'.
But with a simple copy of '\labelformat' from 'figure' to 'subfigure', the output is
figure a page 12
instead of
figure 2.1a page 12
The '\thefigure' counter must be added to 'subfigure' '\labelformat' in order to have the right reference.
It is the same for 'subtable'.