Add the 'mypdf' package
'mypdf' package can insert a external PDF document.
BE CAREFULL, the PDF document can not be insert directly and must be convert in EPS page by page.
The 'Makefile' (or whatever) should be modified in order to take each page of the PDF document and convert it to a new Encapsulated PostScript one-page's document.