Add evolution rate for each analysis
The evolution rate gives a percentage of the evolution for each analysis in order to illustrate the result.
Each R script has been changed (or near each!) to produce a ratio file which could be reuse in the discussion sections.
A 'evolution2tex' has been added in the 'Rinit' file.